updated educations and married_goals

+++ src/educations/educations.controller.ts
@@ -1,4 +1,4 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Request, HttpCode, ValidationPipe } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Request, HttpCode, ValidationPipe, NotFoundException } from '@nestjs/common';
 import { EducationsService } from './educations.service';
 import { CreateEducationDto } from './dto/create-education.dto';
 import { UpdateEducationDto } from './dto/update-education.dto';
@@ -12,10 +12,14 @@ export class EducationsController {
   @Roles(Role.Member)
   @Post()
   create(@Request() req: any, @Body(new ValidationPipe()) data: CreateEducationDto) {
-    return this.educationsService.create(req.user.id, data);
+    try {
+      return this.educationsService.create(req.user.id, data);
+
+    } catch (error) {
+      throw error;
+    }
   }
 
-  //TODO ntar ganti2in pake yang bener
   @Roles(Role.Member)
   @Get()
   findAll(@Request() req: any) {
@@ -26,6 +30,7 @@ export class EducationsController {
       throw error;
     }
   }
+
 
   @Roles(Role.Member)
   @Get(':id')
@@ -53,9 +58,10 @@ export class EducationsController {
 
   @Roles(Role.Member)
   @HttpCode(204)
-  @Delete()
+  @Delete(':id')
   remove(@Request() req: any, @Param('id') id: string) {
     try {
+      if (!id) throw new NotFoundException('Id not found');
       return this.educationsService.remove(req.user.id, id);
 
     } catch (error) {
